refactor: abi for nulls
